Fix: index 자료들의 타입 변경에 따른 .longvalue() 삭제

+++ Back/petPlace/src/main/java/com/minjeok4go/petplace/feed/controller/FeedController.java
@@ -54,7 +54,7 @@ public class FeedController {
                                                                       @AuthenticationPrincipal String tokenUserName
     ) {
         User me = userService.getUserFromToken(tokenUserName);
-        return ResponseEntity.ok(recommendationService.getRecommendedFeeds(me.getId().longValue(), page, size));
+        return ResponseEntity.ok(recommendationService.getRecommendedFeeds(me.getId(), page, size));
     }
 
     @Operation(
